requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-Login
reportlab
Werkzeug
Jinja2<=3.0
itsdangerous==1.1.0
Werkzeug==1.0.1
SQLAlchemy==1.3.23
python-dotenv
